+++ src/types/rabbitmq-config-retry.ts
@@ -3,6 +3,7 @@ import type RabbitMQConfigBase from "src/types/base.config";
 // Тип конфигурации для retry RabbitMQ
 export default interface RabbitMQConfigRetry extends RabbitMQConfigBase {
 	options: RabbitMQConfigBase["options"] & {
+		pattern?: string;
 		queueOptions: {
 			arguments: {
 				"x-dead-letter-exchange": string;